Implemented new subtitles page
Reused the activity, comments, and info tab template code from the
videos page.  Updated those templates a bit.

Updated the comments form view code to work with both the video and
subtitles page.

Still need to implement the javascript to handle the revisions list.

I implemented the form processing code slightly differently then for the
videos page.  This one should work for non-ajax requests as well.  I'd
like to change the video form to work like this as well.

Added way to customize the subtitles page.  This is is pretty much like
how we do the videos page, although there are some slight differences.
I think I'll make the videos page customization more like it when I get
back to working on that page.

Altered how we lookup the team when customizing the subtitles/videos
page.  This one relies on a GET paramater, rather than assuming we want
to use the owner team.  For things like the collab workflow, we also
might want to use an evaluation team, or the work team.

Moved the code to display relative dates from the ActivityRecord class
to a templatetag.