Implemented new team video page
Added some general code for duration/language form fields.  Added
support for the initial selection on the language dropdown.

Added support for a "proxy field".  This is a form input that lives
outside a form, but updates a hidden field inside the form when
it changes.

Renamed assignmentCard, videoCard -> videoCard, videoCardSimple.  We
were using the assignmentCard on the team videos page, so I think this
naming makes more sense.