Standardized forms in the future ui
Wrote template filters to render fields.  Tried to make it so displaying
a field requires as little template code as possible.

Added default labels to the project/video duration fields.

Implemented our own radio select widget so that this can be used rather
than manually rendering it in the template code.