Some form refactoring
Started a UserField class.  This currently only has the backend logic,
still need to implement the UI.

Moved the ui.forms code into a package.  I think we are going to want to
add various widgets and fields classes and I didn't want to stuff all
that code into 1 module.

Refactored the ManagementForm slightly so that it's easier to set it up
in single selection mode.  Added a default permissions_check method.
Created a ManagementFormList class that handles the details of
listing/looking up forms.