Changes to the subtitles step class
This is for pculture/amara-enterprise#874

Allow steps to display a team

Refactored subtitles step action types to make the labels/icons work
similar to django error messages, where you specify a slug and then we
get the label and icon from that.  This makes it easier to test and also
to change up the labels.