Embedder sizing fixes
Take padding into account when sizing the embedder.

Don't initially size the embedder to the width of the video.  This
causes the video to be too wide on the futureui pages for a short time.
It fixes itself when we resize it back to fit the content area, but it's
a noticable delay before that happens.