Implement the delete subtitles form on futureui
Added a new form for this, copied from the one in teams.forms, but
without the obsolete code.  Added some code in the subtitle workflow to
alter the bullet points we display on the form.  Removed the old forms
code.

Started some code to generalize the forms on the videos page.

pculture/amara-enterprise#896