Updated management forms so they can be processed in a worker task
Added ability to pickle forms.  We need this to marshal the data over to
the worker.  Added ability to get progress as the form is processing.

Renamed some attributes, I think these new ones are more clear.

Updated the perform_request method to make it simpler to use.

Added way to get the popular languages in order of popularity