Due date code
Moved the elapsed_time/due_date code out of the template tag code and
into their own module.  We sometimes want to call them from outside the
template system.

Updated the code a bit.  I think this one is cleaner and also fixes some
bugs.  The main one was that the old code was assuming was sometimes
only checking seconds, and not checking days.

Removed the due_date CSS class.  We're not going to use it.