Updated CSS for our step lists
These are the avatars/icons displayed on the assignment history table and the request
management page.  Each avatar/icon represents a step in a request.

Removed the navIcons code.  Replaced it with a new class called
stepList.  Added a styleguide example for stepList.

Updated the default text style for selectList.  The new text looks
better when we display due dates.  Updated the examples.