Hide announcement using javascript
Changed it so the code to hide current announcement is client-side
rather than server side.  This helps with caching, since we can serve
the same content to all anonymous users and let their browsers calculate
if they should see the announcement or not.

Added js.cookie.js library to handle the cookies

Updated announcement HTML to look good for both the old and new styles.