Replaced video listing pages
This includes the videos home page, the featured videos page, and the
latest videos page.  Replaced the old AJAXy code with just a straight
django view version.

Hacked it so we avoid the COUNT() query on video pages.  That was taking
about 1.5 seconds on production to run.

Removed the RPC calls that the javascript code was using.

Updated the CSS a bit.