Updates for the show more link
Replaced "show more" with "show all".  I think if someone clicks that
link they probably would like to see as much as possible, and it's
relatively easy to load all the activity.  This also is more efficient
since before the user might need to click the button a couple times and
we would be constantly reloading the activity.

Added code to show a loading icon while the activity is loading.

Updated the query_string code to work with the new param "show-all"

Put the activity limit in a constant.