Removed browser_id code
This was hurting caching, since we would need to send a different
request for each browser.  I don't think it was accomplishing that much,
since you could still open the editor in two different tabs of the same
browser.  I believe another goal was to allow editing without logging in
to the site, but that's not allowed anymore anyways.