Couple of quick changes to fix fallout from the project filter change
Fixed issue when adding a new video with no project

Made it so we use the project slug in URLs so the current work links for
AoD continue to work.

Only show the project name in as the option label.  This is okay for now
since we're not showing source teams in the ProjectField yet.