Prevent exception with old syncedsubtitleversions (#2770)
I'm not sure how this happens, but we have instances of an multiple
SyncedSubtitleVersion objects in the database for a single
video_url/language tuple.  It seems that this happens when the account
gets deleted, but we don't delete the related SyncedSubtitleVersion
object.  Maybe this has to do with old code, not sure.

In any case, I made it so we figure out which SyncedSubtitleVersion is
the correct one for the current sync account and we use that one.