Changed video URL unique constraint (#2127)
Changed the constraint from (url, type) to (url, team, type).

Now Video.add() won't throw a DuplicateUrl exception if the Video
URL was already added to a different team.

Added some extra handling of new errors that can now happen when we move
a video to a team that already had the video added to it.

Fixed some issues in the videos API when trying to remove a video from a
team.

Removed the unused class teams.oldforms.RemoveTeamVideosForm