Fixed issue with fetching subtitles
The old code had an issue that it was possible for the captions_list API
call to succeed, but then downloading the subtitles failed.  I believe
this happened when you had two accounts linked to a team and the video
was a public video from account that ended up second in the accounts
set.

The new code uses get_video_info to figure out which is the correct
account to try to fetch subtitles from before trying to fetch.