Handle anothr case with prevent_duplicate_public_videos
This code handles the following case:
  - User adds video to the public area
  - User adds video to team A
  - User creates team B with prevent_duplicate_public_videos set
  - User tries to move video from team A -> team B

This should be an error since it would result in the video being in both
the team and the public area.

Added flag to the DuplicateUrlError class to tell if it was caused by
the prevent_duplicate_public_videos.  We want to show different error
messages if it was, and it's not trivial to figure out in the error
handling code.